requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
python-dotenv==0.19.0
scikit-learn==1.2.2
scipy==1.10.1
tqdm==4.65.0
werkzeug==2.0.1
safetensors==0.4.1
accelerate==0.20.3 